testes_acessibilidade_com_Gemini/requirements.txt
flask
google-generativeai
werkzeug
jira
dotenv
google-auth
google-auth-oauthlib
google-api-python-client
google-generativeai